refactor timer logic into a reusable hook
